+++ src/main/java/marvellator/controller/rest/SessionRestController.java
@@ -1,4 +1,4 @@
-package marvellator.controller.apiControllers;
+package marvellator.controller.rest;
 
 
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class SessionApiController {
+public class SessionRestController {
 
     @RequestMapping(value = "api/login", method = RequestMethod.POST)
     public String login() {
